create CounterfactualComponent
